Moved GPG utility to new logic (aligning with SSH utility)
