--- src/functions/gpg/helpers/delete-genkey.js
@@ -1,6 +1,0 @@
-module.exports = (path) => {
-  const fs = require('fs')
-  clog.info('Removing key definition file...')
-  fs.unlinkSync(path)
-  clog.success('File removed!')
-}
