+++ src/functions/gpg/helpers/delete/main.js
@@ -1,0 +1,3 @@
+module.exports = async () => {
+  
+}
