+++ src/functions/gpg/helpers/generate/info-prompter.js
