+++ src/functions/gpg/helpers/generate/user-info.js
