+++ src/functions/gpg/helpers/list/main.js
@@ -1,0 +1,3 @@
+module.exports = async () => {
+  
+}
